newclass script for interfaces

+++ .templates_generators/NewClass.sh
@@ -1,6 +1,9 @@
 if [ -n "$1" ]; then
-	for var in "$@"
-	do
+for var in "$@"
+do
+	if [ "-i" == "$var" ]; then
+		is_interface="true"
+	else
 		echo -e "\033[1;30mGenerating ClassFiles for : $var\033[1;32m";
 		extcpp=".cpp"
 		exthpp=".hpp"
@@ -18,18 +21,22 @@ if [ -n "$1" ]; then
 			after="${after} "
 			shift
 		done
-		sed s/"$filename"/"${a1}${extcpp}${after}"/g ~/.templates_generators/templates/header_42.cpp > ./$a1$extcpp;
-		sed -i.bak s/"{date}"/"$dateTime"/g ./$a1$extcpp >> ./$a1$extcpp;
-		rm ./$a1$extcpp.bak
+		if [ -z "$is_interface" ]; then
+			sed s/"$filename"/"${a1}${extcpp}${after}"/g ~/.templates_generators/templates/header_42.cpp > ./$a1$extcpp;
+			sed -i.bak s/"{date}"/"$dateTime"/g ./$a1$extcpp >> ./$a1$extcpp;
+			rm ./$a1$extcpp.bak
+			sed s/ClassName/$a1/g ~/.templates_generators/templates/class.cpp >> ./$a1$extcpp;
+			sed -i.bak s/"{exthpp}"/"$exthpp"/g ./$a1$extcpp >> ./$a1$extcpp;
+			rm ./$a1$extcpp.bak
+			echo -e "$a1$extcpp created!";
+		fi
 		sed s/"$filename"/"${a1}${exthpp}${after}"/g ~/.templates_generators/templates/header_42.cpp > ./$a1$exthpp;
 		sed -i.bak s/"{date}"/"$dateTime"/g ./$a1$exthpp >> ./$a1$exthpp;
 		rm ./$a1$exthpp.bak
-		sed s/ClassName/$a1/g ~/.templates_generators/templates/class.cpp >> ./$a1$extcpp;
-		sed -i.bak s/"{exthpp}"/"$exthpp"/g ./$a1$extcpp >> ./$a1$extcpp;
-		rm ./$a1$extcpp.bak
 		sed s/ClassName/$a1/g ~/.templates_generators/templates/class.hpp >> ./$a1$exthpp;
-		echo -e "$a1$extcpp and $a1$exthpp created!\033[0;0m";
-	done
+		echo -e "$a1$exthpp created!\033[0;0m";
+	fi
+done
 else
-	echo -e "\033[1;31mUsage :\033[0;0m NewClass [ClassName]";
+	echo -e "\033[1;31mUsage :\033[0;0m NewClass [ClassName]* [-i] [InterfaceName]*";
 fi
